add pagination

+++ voteapp/src/components/index.vue
@@ -1,7 +1,7 @@
-// TODO:
-// add pagination
-// fix the chart
-// user can only edit /delete the poll created by themself
+TODO:
+// add pagination --- done
+// fix the chart --- done
+// user can only edit /delete the poll created by themself --- done
 // add image upload feature
 // share this poll
 
@@ -12,7 +12,7 @@
         <blockquote>
           <h3>Vote Everything
           </h3>
-          <h5>Vote Everything You can vote for anything you want or even better create a new poll!
+          <h5>Vote Everything You can vote for anything you want or even better signup and create a new poll!
           </h5>
         </blockquote>
         <div v-if="loading" class="preloader-wrapper big active">
@@ -35,17 +35,17 @@
         </router-link>
       </div>
       <div class="row">
-        <div class="card" v-for="(poll, index) in polls" :key="index">
+        <div class="card" v-for="(poll, index) in goPage(currentPage)" :key="index">
           <div class="card-image">
             <router-link :to="{name: 'pollDetail', params:{poll_slug:poll.slug}}">
               <img src="@/assets/sample-1.jpg">
             </router-link>
-            <div @click="deletePoll(poll.id)">
+            <div v-if="poll.user_id === loggedUser" @click="deletePoll(poll.id)">
               <i class="material-icons delete right">delete</i>
             </div>
             <span class="card-title">{{poll.title}}</span>
             <router-link :to="{name: 'editPoll', params:{poll_slug: poll.slug}}">
-              <a class="btn-floating halfway-fab waves-effect waves-light red">
+              <a v-if="poll.user_id === loggedUser" class="btn-floating halfway-fab waves-effect waves-light red">
                 <i class="material-icons">edit</i>
               </a>
             </router-link>
@@ -61,25 +61,45 @@
         </div>
       </div>
     </div>
+    <!-- add pagination -->
+    <div v-if="pageNum > 1" id="pagination" class="container pagination center">
+      <ul class="paginationul">
+        <li v-bind:class="{disabled: currentPage == 1, waveEffect:currentPage!= 1}">
+          <a href="javascript:void(0)" @click="switchPage(currentPage-1)">
+            <i class="material-icons">chevron_left</i>
+          </a>
+        </li>
+        <li v-for="(page, index) in pageNum" :key="index" v-bind:class="{active: currentPage == page}" class="waves-effect">
+          <a href="javascript:void(0)" @click="switchPage(page)">{{page}}</a>
+        </li>
+        <li v-bind:class="{disabled: currentPage == pageNum, waveEffect:currentPage!= pageNum}">
+          <a href="javascript:void(0)" @click="switchPage(currentPage+1)">
+            <i class="material-icons">chevron_right</i>
+          </a>
+        </li>
+      </ul>
+    </div>
   </div>
-  <!-- add pagination -->
 
 </template>
 
 <script>
 import db from "@/firebase/init";
 import { bus } from "../main.js";
+import firebase from "firebase";
 
 export default {
   data() {
     return {
       polls: [],
-      loading: true
+      loading: true,
+      pageSize: 9,
+      pageNum: 1,
+      currentPage: 1
     };
   },
   created() {
     bus.$on("searchPoll", data => {
-      console.log(data);
       this.polls = this.polls.filter(poll => {
         return poll.title.match(data.toString);
       });
@@ -96,17 +116,41 @@ export default {
         });
       });
   },
+  computed: {
+    loggedUser() {
+      let user = firebase.auth().currentUser;
+      if (user) {
+        return user.uid;
+      } else {
+        return null;
+      }
+    }
+  },
   methods: {
     deletePoll(id) {
-      db
-        .collection("polls")
-        .doc(id)
-        .delete()
-        .then(() => {
-          this.polls = this.polls.filter(filterPoll => {
-            return filterPoll.id != id;
+      if (firebase.auth().currentUser) {
+        db
+          .collection("polls")
+          .doc(id)
+          .delete()
+          .then(() => {
+            this.polls = this.polls.filter(filterPoll => {
+              return filterPoll.id != id;
+            });
           });
-        });
+      }
+    },
+    switchPage(num) {
+      if (num <= this.pageNum && num >= 1) {
+        this.currentPage = num;
+      }
+    },
+    goPage(num) {
+      this.pageNum = Math.ceil(this.polls.length / this.pageSize);
+      this.currentPage = num;
+      let start = (this.currentPage - 1) * this.pageSize;
+      let end = start + this.pageSize + 1;
+      return this.polls.slice(start, end);
     }
   }
 };
@@ -133,5 +177,8 @@ export default {
   top: 4px;
   right: 4px;
   cursor: pointer;
+}
+#pagination {
+  margin-bottom: 60px;
 }
 </style>
